C code to return last digit and print it

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -8,7 +8,7 @@
  */
 int print_last_digit(int n)
 {
-	long int last;
+	unsigned int last;
 
 	if (n < 0)
 		n = n * -1;
